+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Test: Upgrade nim-waku node to v0.4.
+
 ## [0.6.0] - 2021-06-09
 
 ### Changed