+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Test: Upgrade nim-waku node to v0.4.
 - Waku Light Push upgraded to `2.0.0-beta1`.
+- Examples (web chat): Catch error if chat message decoding fails.
 
 ## [0.6.0] - 2021-06-09